work in progress

+++ apps/health_campaign_field_worker_app/lib/models/performance_summary.dart
@@ -6,8 +6,8 @@ class PerformanceSummary extends EntityModel {
   double treatedPercentage;
   int householdCount;
   int taskCount;
-  int drugOne;
-  int drugTwo;
+  double drugOne;
+  double drugTwo;
 
   PerformanceSummary({
     required this.treatedPercentage,
